Rustdoc examples for Redis
