+++ src/lib.rs
@@ -60,41 +60,8 @@ pub mod http;
 #[allow(missing_docs)]
 pub mod mqtt;
 
-/// Redis storage and messaging.
 #[allow(missing_docs)]
-pub mod redis {
-    use std::hash::{Hash, Hasher};
-
-    pub use super::wit::v2::redis::{Connection, Error, Payload, RedisParameter, RedisResult};
-
-    impl PartialEq for RedisResult {
-        fn eq(&self, other: &Self) -> bool {
-            use RedisResult::*;
-            match (self, other) {
-                (Nil, Nil) => true,
-                (Status(a), Status(b)) => a == b,
-                (Int64(a), Int64(b)) => a == b,
-                (Binary(a), Binary(b)) => a == b,
-                _ => false,
-            }
-        }
-    }
-
-    impl Eq for RedisResult {}
-
-    impl Hash for RedisResult {
-        fn hash<H: Hasher>(&self, state: &mut H) {
-            use RedisResult::*;
-
-            match self {
-                Nil => (),
-                Status(s) => s.hash(state),
-                Int64(v) => v.hash(state),
-                Binary(v) => v.hash(state),
-            }
-        }
-    }
-}
+pub mod redis;
 
 pub mod pg;
